requirements.txt
pyelftools==0.30
PyGithub==2.2.0
PyGithub==2.2.0
Requests==2.31.0
